src/composite/pyproject.toml
[project]
name = "composite"
version = "0.1.0"
description = "MCP composite server for aggregating multiple backend MCP servers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser",
    "docx",
    "langquery",
    "pdf",
    "pptx",
    "pyyaml>=6.0",
    "starlette>=0.41.3",
    "vectorstore",
    "xlsx",
]

[tool.uv.sources]
browser = { workspace = true }
docx = { workspace = true }
langquery = { workspace = true }
pdf = { workspace = true }
pptx = { workspace = true }
vectorstore = { workspace = true }
xlsx = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
